requirements.txt
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.2.0
bleach==4.1.0
build==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.13
docutils==0.18.1
frozenlist==1.2.0
gitdb==4.0.9
GitPython==3.1.24
idna==3.3
importlib-metadata==4.10.0
keyring==23.4.0
multidict==5.2.0
packaging==21.3
pep517==0.12.0
pkginfo==1.8.2
pycparser==2.21
PyGithub==1.55
Pygments==2.10.0
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==3.0.6
readme-renderer==32.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==10.13.0
six==1.16.0
smmap==5.0.0
tomli==2.0.0
tqdm==4.62.3
twine==3.7.1
typing-extensions==3.10.0.2
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
